+++ src/giirs/giirs_nc2ioda.py
@@ -37,43 +37,12 @@ AttrData = DefaultOrderedDict(lambda: DefaultOrderedDict(dict))
 
 
 class LwRadiance(object):
-    def __init__(self, filenames, writer):
-        self.filenames = filenames        
+    def __init__(self, writer):
         self.data = DefaultOrderedDict(lambda: DefaultOrderedDict(dict))
         self.units_values = {}  # cannot use an ordered dict for this
         self.writer = writer
-        self.readInFiles()
 
-    def xferMdataVar(self, ncd, inName, outName, mData, dType):
-        # This method adds a variable to the given metadata dictionary element.
-        ncVar = ncd.variables[inName]
-        mData[outName] = self.writer.FillNcVector(ncVar[:], dType)
-        if 'units' in ncVar.ncattrs():
-            self.units_values[outName] = ncVar.getncattr('units')
-
-    def xferLocMdataVar(self, ncd, numFiles, inName, outName, mData, dType,mask):
-        # This method adds a variable to the given metadata dictionary element.
-        ncVar = ncd.variables[inName]
-        #print(inName, len(ncVar[mask]))
-        if (numFiles == 1):
-            mData[outName] = self.writer.FillNcVector(ncVar[mask], dType)
-            if 'units' in ncVar.ncattrs():
-                self.units_values[outName] = ncVar.getncattr('units')
-        else:
-            mData[outName] = np.append(mData[outName], self.writer.FillNcVector(ncVar[mask], dType), axis=0)
-            
-    def xferLocMdataVar_detector(self, ncd, numFiles, detector, outName, mData, dType,mask):
-        # This method adds a variable to the given metadata dictionary element.
-        ncVar = detector
-        if (numFiles == 1):
-            mData[outName] = self.writer.FillNcVector(ncVar[mask], dType)
-            self.units_values[outName] = "1"
-        else:
-            mData[outName] = np.append(mData[outName], self.writer.FillNcVector(ncVar[mask], dType), axis=0)
-            
-        
-            
-    def readInFiles(self):
+    def readInFiles(self, filenames):
         # This method will transfer variable data from the input file into a
         # dictionary (self.data) that the netcdf writer understands.
         #
@@ -112,133 +81,165 @@ class LwRadiance(object):
 
         numFiles = 0
         self.writer._nlocs = 0
-        giirsfiletxt = open(self.filenames[0], 'r')
-        giirsfiles   = giirsfiletxt.readlines()
+        with open(self.filenames[0], 'r') as giirsfiletxt:
+            giirsfiles   = giirsfiletxt.readlines()
 
-        giirsfiletxt.close()
         sat_zen_max = 74
         detector_number = 128
-        detector_array  = np.asarray ( range (1, detector_number+1) )
+        detector_array = np.arange(1, detector_number+1,dtype='i4')
         
         #wavenumber could also read from giirs L1 data file. 
         wn_lw=np.arange(700,1130.001,0.625)
         wn_mw=np.arange(1650,2250.001,0.625)        
-        for ff in giirsfiles:
-            
-            f=ff.strip()
-            ncd = nc.Dataset(f, 'r')
-            print(f)
-            sat_zen = ncd.variables['IRLW_SatelliteZenith'][:]
-            mask = (sat_zen < sat_zen_max )      
-            nlocs = len(sat_zen[mask])
-            print("Valid obs number: ",nlocs)
-            if nlocs == 0:
-                continue    
-                
-            numFiles += 1
-            
-            nchans = ncd.dimensions["LWchannel"].size
-            self.writer._nlocs += nlocs
-            self.writer._nvars = nchans
-            self.writer._nrecs = 1
+        for f in (ff.strip() for ff in giirsfiles):
+            with nc.Dataset(f, 'r') as ncd:
+                print("[GIIRS2IODA] Processing: %s"%f)
+                sat_zen = ncd.variables['IRLW_SatelliteZenith'][:]
+                mask = (sat_zen < sat_zen_max )
+                nlocs = len(sat_zen[mask])
+                if nlocs == 0:
+                    print("[GIIRS2IODA] NO VALID OBS. Skipping.")
+                    continue
+                print("[GIIRS2IODA] #Valid obs: %d"%nlocs)
 
-            # Channel metadata associated with longwave radiance
-            if (numFiles == 1):
-                self.xferMdataVar(ncd, 'LW_wnum', 'channel_wavenumber', VarMdata, 'float')
+                numFiles += 1
 
-                VarMdata['channel_number'] = self.writer.FillNcVector(
-                    np.ma.array(range(1, nchans+1)), 'integer')
-                self.units_values['channel_number'] = '1'
+                nchans = ncd.dimensions["LWchannel"].size
+                self.writer._nlocs += nlocs
+                self.writer._nvars = nchans
+                self.writer._nrecs = 1
 
-            # The file contains a single image with a sub-second scan time. We
-            # are currently retaining date-time stamps to the nearest second so
-            # for now just grab the beginning scan time and use that for all locations.
-            # The variables holding the time offset from 1 Jan, 0000 are problematic
-            # since the python datetime package doesn't allow year zero, and it's not
-            # clear how days are coverted over several millinnea. The observation
-            # date and time are also in global attributes and appear to line up with
-            # the file name.
-            obsDate = ncd.getncattr("Observing Beginning Date").split("-")
-            obsTime = ncd.getncattr("Observing Beginning Time").split(":")
-            obsSeconds = timedelta(seconds=int(round(float(obsTime[2]))))
-            obsDateTime = datetime(year=int(obsDate[0]), month=int(obsDate[1]),
-                                   day=int(obsDate[2]), hour=int(obsTime[0]),
-                                   minute=int(obsTime[1])) + obsSeconds
-            obsDtimeString = obsDateTime.strftime("%Y-%m-%dT%H:%M:%SZ")
-
-            # Form a vector nlocs long containing the datetime stamp
-            if (numFiles == 1):
-                LocMdata['datetime'] = self.writer.FillNcVector(
-                    np.full((nlocs), obsDtimeString), 'datetime')
-                self.units_values['datetime'] = 'ISO 8601 format'
-            else:
-                LocMdata['datetime'] = np.append(LocMdata['datetime'], self.writer.FillNcVector(np.full((nlocs), obsDtimeString), 'datetime'), axis=0)
-
-            # Read in the latitude and longitude associated with the long wave data
-            self.xferLocMdataVar(ncd, numFiles, 'IRLW_Latitude', 'latitude', LocMdata, 'float', mask)
-            self.xferLocMdataVar(ncd, numFiles, 'IRLW_Longitude', 'longitude', LocMdata, 'float', mask)
-            #=====detector number 
-            self.xferLocMdataVar_detector(ncd, numFiles, detector_array*1.0, 'scan_position', LocMdata, 'float', mask)
-
-            # Read in the instrument meta data associated with the long wave data
-            self.xferLocMdataVar(ncd, numFiles, 'IRLW_SolarZenith', 'solar_zenith_angle', LocMdata, 'float', mask)
-            self.xferLocMdataVar(ncd, numFiles, 'IRLW_SolarAzimuth', 'solar_azimuth_angle', LocMdata, 'float', mask)
-            self.xferLocMdataVar(ncd, numFiles, 'IRLW_SatelliteZenith', 'sensor_zenith_angle', LocMdata, 'float', mask)
-            self.xferLocMdataVar(ncd, numFiles, 'IRLW_SatelliteZenith', 'sensor_view_angle', LocMdata, 'float', mask)
-            self.xferLocMdataVar(ncd, numFiles, 'IRLW_SatelliteAzimuth', 'sensor_azimuth_angle', LocMdata, 'float', mask)
-            
-            self.xferLocMdataVar(ncd, numFiles, 'Cloud_Fraction', 'cloud_fraction', LocMdata, 'float', mask)
-
-            # Read in the long wave radiance
-            # For now fabricate the QC marks and error estimates
-            ncVar = ncd.variables['ES_RealLW']
-            lwRad = ncVar[:,mask]
-            print(self.writer._nlocs)
-            H  = 6.6237E-27
-            C  = 2.99791E+10
-            B  = 1.38024E-16
-            C1 = 2.*H*C*C
-            C2 = H*C/B
-            # GIIRS wavenumbers
-
-            TBB = lwRad*0-999
-            
-            for iloc in range(nlocs):            
-                for ivar in range(1,nchans-1):
-                #Hamming apodization for Ch2 - Ch688
-                    lwRadiance_apo = 0.23*lwRad[ivar-1,iloc] +0.54*lwRad[ivar,iloc] + 0.23*lwRad[ivar+1,iloc]
-                #Radiane to Bright Temperature  
-                    if ( (lwRadiance_apo > 0) and (lwRadiance_apo < 300) ):
-                        TBB[ivar,iloc] = planck.rad2bt( wn_lw[ivar]*100, lwRadiance_apo/pow(10,5) )
-                #For the first channel Ch1 and last channel Ch689 in long wave band
-                ivar = 0
-                lwRadiance_apo = lwRad[ivar,iloc]
-                if ( (lwRadiance_apo > 0) and (lwRadiance_apo < 300) ):
-                    TBB[ivar,iloc] = planck.rad2bt(wn_lw[ivar]*100, lwRadiance_apo/pow(10,5))
-                
-                ivar = nchans-1
-                lwRadiance_apo = lwRad[ivar,iloc]
-                if ( (lwRadiance_apo > 0) and (lwRadiance_apo < 300) ):
-                    TBB[ivar,iloc] = planck.rad2bt(wn_lw[ivar]*100, lwRadiance_apo/pow(10,5))
-            
-            if 'units' in ncVar.ncattrs():
-                Units = ncVar.getncattr('units')
-            else:
-                Units = None
-            for ivar in range(nchans):
-                varName = "brightness_temperature_%d" % (ivar + 1)
+                # Channel metadata associated with longwave radiance
                 if (numFiles == 1):
-                    self.data[(varName, 'ObsValue')] = self.writer.FillNcVector(TBB[ivar, :], 'float')
-                    self.data[(varName, 'PreQC')] = self.writer.FillNcVector(np.full((nlocs), 0), 'integer')
-                    self.data[(varName, 'ObsError')] = self.writer.FillNcVector(np.full((nlocs), 2.0), 'float')
-                else:
-                    self.data[(varName, 'ObsValue')] = np.append(self.data[(varName, 'ObsValue')], self.writer.FillNcVector(TBB[ivar, :], 'float'), axis=0)
-                    self.data[(varName, 'PreQC')] = np.append(self.data[(varName, 'PreQC')], self.writer.FillNcVector(np.full((nlocs), 0), 'integer'), axis=0)
-                    self.data[(varName, 'ObsError')] = np.append(self.data[(varName, 'ObsError')], self.writer.FillNcVector(np.full((nlocs), 2.0), 'float'), axis=0)
-                if Units:
-                   self.units_values[varName] = 'K'
+                    self.xferMdataVar(ncd, 'LW_wnum', 'channel_wavenumber', VarMdata, 'float')
 
-            ncd.close()
+                    VarMdata['channel_number'] = self.writer.FillNcVector(
+                        np.ma.array(range(1, nchans+1)), 'integer')
+                    self.units_values['channel_number'] = '1'
+
+                # The file contains a single image with a sub-second scan time. We
+                # are currently retaining date-time stamps to the nearest second so
+                # for now just grab the beginning scan time and use that for all locations.
+                # The variables holding the time offset from 1 Jan, 0000 are problematic
+                # since the python datetime package doesn't allow year zero, and it's not
+                # clear how days are coverted over several millinnea. The observation
+                # date and time are also in global attributes and appear to line up with
+                # the file name.
+                obsDate = ncd.getncattr("Observing Beginning Date").split("-")
+                obsTime = ncd.getncattr("Observing Beginning Time").split(":")
+                obsSeconds = timedelta(seconds=int(round(float(obsTime[2]))))
+                obsDateTime = datetime(year=int(obsDate[0]), month=int(obsDate[1]),
+                                    day=int(obsDate[2]), hour=int(obsTime[0]),
+                                    minute=int(obsTime[1])) + obsSeconds
+                obsDtimeString = obsDateTime.strftime("%Y-%m-%dT%H:%M:%SZ")
+
+                # Form a vector nlocs long containing the datetime stamp
+                if (numFiles == 1):
+                    LocMdata['datetime'] = self.writer.FillNcVector(
+                        np.full((nlocs), obsDtimeString), 'datetime')
+                    self.units_values['datetime'] = 'ISO 8601 format'
+                else:
+                    LocMdata['datetime'] = np.append(LocMdata['datetime'], self.writer.FillNcVector(np.full((nlocs), obsDtimeString), 'datetime'), axis=0)
+
+                # Read in the latitude and longitude associated with the long wave data
+                self.xferLocMdataVar(ncd, numFiles, 'IRLW_Latitude', 'latitude', LocMdata, 'float', mask)
+                self.xferLocMdataVar(ncd, numFiles, 'IRLW_Longitude', 'longitude', LocMdata, 'float', mask)
+                #=====detector number
+                self.xferLocMdataVar_detector(ncd, numFiles, detector_array*1.0, 'scan_position', LocMdata, 'float', mask)
+
+                # Read in the instrument meta data associated with the long wave data
+                self.xferLocMdataVar(ncd, numFiles, 'IRLW_SolarZenith', 'solar_zenith_angle', LocMdata, 'float', mask)
+                self.xferLocMdataVar(ncd, numFiles, 'IRLW_SolarAzimuth', 'solar_azimuth_angle', LocMdata, 'float', mask)
+                self.xferLocMdataVar(ncd, numFiles, 'IRLW_SatelliteZenith', 'sensor_zenith_angle', LocMdata, 'float', mask)
+                self.xferLocMdataVar(ncd, numFiles, 'IRLW_SatelliteZenith', 'sensor_view_angle', LocMdata, 'float', mask)
+                self.xferLocMdataVar(ncd, numFiles, 'IRLW_SatelliteAzimuth', 'sensor_azimuth_angle', LocMdata, 'float', mask)
+
+                self.xferLocMdataVar(ncd, numFiles, 'Cloud_Fraction', 'cloud_fraction', LocMdata, 'float', mask)
+
+                # Read in the long wave radiance
+                # For now fabricate the QC marks and error estimates
+                ncVar = ncd.variables['ES_RealLW']
+                lwRad = ncVar[:,mask]
+                print("[GIIRS2IODA] #Locs: %d"%self.writer._nlocs)
+
+                # GIIRS wavenumbers
+                TBB = lwRad*0-999
+
+                for iloc in range(nlocs):
+                    for ivar in range(1,nchans-1):
+                    #Hamming apodization for Ch2 - Ch688
+                        lwRadiance_apo = 0.23*lwRad[ivar-1,iloc] +0.54*lwRad[ivar,iloc] + 0.23*lwRad[ivar+1,iloc]
+                    #Radiane to Bright Temperature
+                        if ( (lwRadiance_apo > 0) and (lwRadiance_apo < 300) ):
+                            TBB[ivar,iloc] = self.rad2bt( wn_lw[ivar]*100, lwRadiance_apo/pow(10,5) )
+                    #For the first channel Ch1 and last channel Ch689 in long wave band
+                    ivar = 0
+                    lwRadiance_apo = lwRad[ivar,iloc]
+                    if ( (lwRadiance_apo > 0) and (lwRadiance_apo < 300) ):
+                        TBB[ivar,iloc] = self.rad2bt(wn_lw[ivar]*100, lwRadiance_apo/pow(10,5))
+
+                    ivar = nchans-1
+                    lwRadiance_apo = lwRad[ivar,iloc]
+                    if ( (lwRadiance_apo > 0) and (lwRadiance_apo < 300) ):
+                        TBB[ivar,iloc] = self.rad2bt(wn_lw[ivar]*100, lwRadiance_apo/pow(10,5))
+
+                if 'units' in ncVar.ncattrs():
+                    Units = ncVar.getncattr('units')
+                else:
+                    Units = None
+                for ivar in range(nchans):
+                    varName = "brightness_temperature_%d" % (ivar + 1)
+                    if (numFiles == 1):
+                        self.data[(varName, 'ObsValue')] = self.writer.FillNcVector(TBB[ivar, :], 'float')
+                        self.data[(varName, 'PreQC')] = self.writer.FillNcVector(np.full((nlocs), 0), 'integer')
+                        self.data[(varName, 'ObsError')] = self.writer.FillNcVector(np.full((nlocs), 2.0), 'float')
+                    else:
+                        self.data[(varName, 'ObsValue')] = np.append(self.data[(varName, 'ObsValue')], self.writer.FillNcVector(TBB[ivar, :], 'float'), axis=0)
+                        self.data[(varName, 'PreQC')] = np.append(self.data[(varName, 'PreQC')], self.writer.FillNcVector(np.full((nlocs), 0), 'integer'), axis=0)
+                        self.data[(varName, 'ObsError')] = np.append(self.data[(varName, 'ObsError')], self.writer.FillNcVector(np.full((nlocs), 2.0), 'float'), axis=0)
+                    if Units:
+                        self.units_values[varName] = 'K'
+
+    ## Helper Methods
+
+    def xferMdataVar(self, ncd, inName, outName, mData, dType):
+        # This method adds a variable to the given metadata dictionary element.
+        ncVar = ncd.variables[inName]
+        mData[outName] = self.writer.FillNcVector(ncVar[:], dType)
+        if 'units' in ncVar.ncattrs():
+            self.units_values[outName] = ncVar.getncattr('units')
+
+    def xferLocMdataVar(self, ncd, numFiles, inName, outName, mData, dType,mask):
+        # This method adds a variable to the given metadata dictionary element.
+        ncVar = ncd.variables[inName]
+        #print(inName, len(ncVar[mask]))
+        if (numFiles == 1):
+            mData[outName] = self.writer.FillNcVector(ncVar[mask], dType)
+            if 'units' in ncVar.ncattrs():
+                self.units_values[outName] = ncVar.getncattr('units')
+        else:
+            mData[outName] = np.append(mData[outName], self.writer.FillNcVector(ncVar[mask], dType), axis=0)
+
+    def xferLocMdataVar_detector(self, ncd, numFiles, detector, outName, mData, dType,mask):
+        # This method adds a variable to the given metadata dictionary element.
+        ncVar = detector
+        if (numFiles == 1):
+            mData[outName] = self.writer.FillNcVector(ncVar[mask], dType)
+            self.units_values[outName] = "1"
+        else:
+            mData[outName] = np.append(mData[outName], self.writer.FillNcVector(ncVar[mask], dType), axis=0)
+
+    @classmethod
+    def rad2bt(wavnum, radiance):
+
+        H_PLANCK = 6.62606957 * 1e-34  # SI-unit = [J*s]
+        K_BOLTZMANN = 1.3806488 * 1e-23  # SI-unit = [J/K]
+        C_SPEED = 2.99792458 * 1e8  # SI-unit = [m/s]
+
+        const1 = H_PLANCK * C_SPEED / K_BOLTZMANN
+        const2 = 2 * H_PLANCK * C_SPEED**2
+        bt = const1 * wavnum / np.log(np.divide(const2 * wavnum**3, radiance) + 1.0)
+        return bt
 
 
 def main():
@@ -267,7 +268,8 @@ def main():
     writer = iconv.NcWriter(args.output, [], locationKeyList)
 
     # Read in the giirs lw radiance
-    lwrad = LwRadiance(args.input, writer)
+    lwrad = LwRadiance(writer)
+    lwrad.readInFiles(args.input)
 
     # Add the 'date_time_string' attribute which sets the reference datetime for
     # the observations.
